chore: rename to throttle
Unfortunately it seems, that go really wants to build the executables
based on the project name. So after renaming the python version to
throttle-cli (it had a conflict anyway so the pypi package is already
throttle-cli) rename this project to throttle.

Note, maybe I should come up with a better name.